Let the compiler know that SULog is a format function.
This allows it to warn properly on malformed format strings.

+++ SUDiskImageUnarchiver.m
@@ -80,7 +80,7 @@
 		if (taskResult != 0)
 		{
 			NSString*	resultStr = output ? [[[NSString alloc] initWithData: output encoding: NSUTF8StringEncoding] autorelease] : nil;
-			SULog( @"hdiutil failed with code: %d data: <<%@>>", taskResult, resultStr );
+			SULog( @"hdiutil failed with code: %ld data: <<%@>>", (long)taskResult, resultStr );
 			goto reportError;
 		}
 		mountedSuccessfully = YES;
